Complete documentation
also fixed SCROLL_LEFT & SCROLL_RIGHT macros names' to match the motion of the bricks instead of that of the field, as the manual contextualises it as the former, and the latter makes little sense when scrolling is in different directions for different rows (i.e. every scroller except stage 02).